[Aman] : Added notification service, andmin controller and admin pannel and fix css in notification and booking, also added configutration of rabbitmq

+++ parkinglot-service/src/main/java/com/parkease/parkinglot_service/dtos/ParkingLotRequestDto.java
@@ -37,7 +37,7 @@ public class ParkingLotRequestDto {
 
     @NotNull(message = "Manager id is required")
     @Positive(message = "Manager id must be greater than 0")
-    private Integer managerId;
+    private Long managerId;
 
     @NotNull(message = "Open status is required")
     private Boolean open;
